Optimize tracing by skipping a lot of entities
Actually do what the comment says and "limit it to the part of the move
not already clipped off by the world".

The commented-out line which is deleted would have been wrong because
changing the endpoint would result in wrong results for fraction and
allsolid. But using the reduced endpoint for the entity search area is
fine.